+++ apps/frontend/src/libs/hooks/hooks.ts
@@ -1,6 +1,16 @@
 export { useAppDispatch } from "./use-app-dispatch/use-app-dispatch.hook.js";
 export { useAppForm } from "./use-app-form/use-app-form.hook.js";
 export { useAppSelector } from "./use-app-selector/use-app-selector.hook.js";
-export { useCallback, useEffect, useMemo, useRef, useState } from "react";
+export { useMapClient } from "./use-map-client/use-map-client.hook.js";
+export { useTabNavigation } from "./use-tab-navigation/use-tab-navigation.hook.js";
+export {
+	createContext,
+	useCallback,
+	useContext,
+	useEffect,
+	useMemo,
+	useRef,
+	useState,
+} from "react";
 export { useController as useFormController } from "react-hook-form";
-export { useLocation } from "react-router";
+export { useLocation, useSearchParams } from "react-router";