+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -1,5 +1,6 @@
 type IconName =
 	| "analytics"
+	| "arrowDown"
 	| "close"
 	| "dashboard"
 	| "map"