+++ apps/frontend/src/modules/auth/slices/auth.slice.ts
@@ -1,19 +1,27 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { type UserAuthResponseDto } from "@smartscapes/shared";
 
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
+import { type UserAuthResponseDto } from "~/modules/users/users.js";
 
-import { getAuthenticatedUser, signIn, signUp } from "./actions.js";
+import {
+	getAuthenticatedUser,
+	logout,
+	patchAuthenticatedUser,
+	signIn,
+	signUp,
+} from "./actions.js";
 
 type State = {
 	authenticatedUser: null | UserAuthResponseDto;
+	authenticatedUserPatchStatus: ValueOf<typeof DataStatus>;
 	dataStatus: ValueOf<typeof DataStatus>;
 	isInitialized: boolean;
 };
 
 const initialState: State = {
 	authenticatedUser: null,
+	authenticatedUserPatchStatus: DataStatus.IDLE,
 	dataStatus: DataStatus.IDLE,
 	isInitialized: false,
 };
@@ -32,6 +40,11 @@ const { actions, name, reducer } = createSlice({
 			state.authenticatedUser = null;
 			state.dataStatus = DataStatus.REJECTED;
 			state.isInitialized = true;
+		});
+
+		builder.addCase(logout.fulfilled, (state) => {
+			state.authenticatedUser = null;
+			state.dataStatus = DataStatus.FULFILLED;
 		});
 
 		builder.addCase(signUp.fulfilled, (state, action) => {
@@ -56,6 +69,18 @@ const { actions, name, reducer } = createSlice({
 		builder.addCase(signIn.rejected, (state) => {
 			state.authenticatedUser = null;
 			state.dataStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(patchAuthenticatedUser.fulfilled, (state, action) => {
+			state.authenticatedUser = action.payload.data;
+			state.authenticatedUserPatchStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(patchAuthenticatedUser.pending, (state) => {
+			state.authenticatedUserPatchStatus = DataStatus.PENDING;
+		});
+		builder.addCase(patchAuthenticatedUser.rejected, (state) => {
+			state.authenticatedUser = null;
+			state.authenticatedUserPatchStatus = DataStatus.REJECTED;
 		});
 	},
 	initialState,