+++ apps/frontend/src/modules/auth/slices/auth.ts
@@ -1,9 +1,17 @@
-import { getAuthenticatedUser, signIn, signUp } from "./actions.js";
+import {
+	getAuthenticatedUser,
+	logout,
+	patchAuthenticatedUser,
+	signIn,
+	signUp,
+} from "./actions.js";
 import { actions } from "./auth.slice.js";
 
 const allActions = {
 	...actions,
 	getAuthenticatedUser,
+	logout,
+	patchAuthenticatedUser,
 	signIn,
 	signUp,
 };