+++ packages/shared/src/modules/user-follows/libs/types/user-follows-parameters-dto.type.ts
@@ -1,0 +1,5 @@
+type UserFollowsParametersDto = {
+	userId: string;
+};
+
+export { type UserFollowsParametersDto };